requirements.txt
langchain==0.3.23
langchain-community==0.3.21
langchain-core==0.3.56
langchain-openai==0.3.15
langchain-pinecone==0.2.5
langchain-tests==0.3.19
langchain-text-splitters==0.3.8
langsmith==0.3.39
livekit==1.0.6
livekit-agents==1.0.18
livekit-api==1.0.2
livekit-plugins-cartesia==1.0.18
livekit-plugins-deepgram==1.0.18
livekit-plugins-openai==1.0.18
livekit-plugins-silero==1.0.18
livekit-plugins-turn-detector==1.0.18
livekit-protocol==1.0.2
MarkupSafe==3.0.2
marshmallow==3.26.1
openai==1.76.2
pinecone==6.0.2
pinecone-plugin-interface==0.0.7
regex==2024.11.6
requests==2.32.3
tiktoken==0.9.0
tokenizers==0.21.1
tqdm==4.67.1
urllib3==2.4.0
aiohttp>=3.10,<3.11
python-dotenv
livekit-plugins-noise-cancellation